refactor(metadata): Extract wallet operations to wallet.rs
Part of #51 - Extract ~21 wallet-related database operations to
a dedicated wallet.rs module following the pattern established
in PRs #73 (pool.rs) and #74 (user.rs).

Methods extracted:
- Wallet CRUD: descriptor_exists, extract_checksum, insert_wallet,
  get_wallet_by_descriptor, get_wallet_by_checksum, update_wallet_by_checksum,
  mark_wallet_as_deleted, hard_delete_wallet_by_checksum
- Wallet queries: get_all_wallets, get_wallets_for_user,
  get_wallets_for_user_oldest_first, count_wallets_for_user,
  is_wallet_owned_by_user, get_wallets_for_tier_sync,
  get_non_syncing_wallets_summary, get_ready_wallets, get_deleted_wallets
- Wallet status: update_wallet_balance_by_checksum, update_wallet_last_synced,
  update_wallet_status, update_wallet_active_status

Closes #55
