+++ backend/src/metadata/mod.rs
@@ -2,6 +2,7 @@ mod db;
 mod pool;
 mod types;
 mod user;
+mod wallet;
 
 pub use pool::MetadataDb;
 pub use types::*;
