+++ backend/src/metadata/wallet.rs
@@ -1,0 +1,724 @@
+use super::pool::MetadataDb;
+use super::types::*;
+use anyhow::{anyhow, Result};
+use bdk_wallet::rusqlite::{params, ToSql};
+use tokio::task::spawn_blocking;
+
+impl MetadataDb {
+    // ============================
+    // WALLET CRUD OPERATIONS
+    // ============================
+
+    pub async fn descriptor_exists(&self, descriptor: &str) -> Result<bool> {
+        let pool = self.pool.clone();
+        let descriptor = descriptor.to_string();
+
+        spawn_blocking(move || -> Result<bool> {
+            let conn = pool.get()?;
+            let count: i64 = conn.query_row(
+                "SELECT COUNT(*) FROM wallets WHERE descriptor = ?1",
+                params![descriptor],
+                |row| row.get(0),
+            )?;
+            Ok(count > 0)
+        })
+        .await?
+    }
+
+    /// Extract checksum from a Bitcoin descriptor
+    pub fn extract_checksum(&self, descriptor: &str) -> String {
+        extract_checksum(descriptor)
+    }
+
+    pub async fn insert_wallet(
+        &self,
+        name: &str,
+        descriptor: &str,
+        user_id: &str,
+    ) -> Result<String> {
+        let pool = self.pool.clone();
+        let name = name.to_string();
+        let descriptor = descriptor.to_string();
+        let user_id = user_id.to_string();
+
+        spawn_blocking(move || -> Result<String> {
+            let conn = pool.get()?;
+            let hex_color = calculate_wallet_color(&descriptor);
+            let current_time = chrono::Utc::now().format("%Y-%m-%d %H:%M:%S").to_string();
+
+            // Extract checksum from descriptor (part after #)
+            let checksum = descriptor
+                .split('#')
+                .next_back()
+                .ok_or_else(|| anyhow::anyhow!("Invalid descriptor format: missing checksum"))?
+                .to_string();
+
+            conn.execute(
+                "INSERT INTO wallets (checksum, name, descriptor, hex_color, balance_total, last_activity, status, user_id) VALUES (?1, ?2, ?3, ?4, ?5, ?6, ?7, ?8)",
+                params![&checksum, &name, &descriptor, &hex_color, "0", &current_time, "pending", user_id],
+            )?;
+            Ok(checksum)
+        })
+        .await?
+    }
+
+    pub async fn get_wallet_by_descriptor(
+        &self,
+        descriptor: &str,
+    ) -> Result<Option<WalletMetadata>> {
+        let pool = self.pool.clone();
+        let descriptor = descriptor.to_string();
+
+        spawn_blocking(move || -> Result<Option<WalletMetadata>> {
+            let conn = pool.get()?;
+            match conn.query_row(
+                "SELECT w.checksum, w.name, w.descriptor, w.hex_color, w.created_at, w.balance_total,
+                        (SELECT MAX(t.first_seen_at) FROM transactions t WHERE t.wallet_checksum = w.checksum) as last_activity,
+                        w.status, COUNT(c.id) as contact_count, w.user_id, w.is_active
+                 FROM wallets w
+                 LEFT JOIN contacts c ON w.checksum = c.wallet_checksum
+                 WHERE w.descriptor = ?1
+                 GROUP BY w.checksum, w.name, w.descriptor, w.hex_color, w.created_at, w.balance_total, w.status, w.user_id, w.is_active",
+                params![descriptor],
+                |row| {
+                    Ok(WalletMetadata {
+                        checksum: row.get(0)?,
+                        name: row.get(1)?,
+                        descriptor: row.get(2)?,
+                        hex_color: row.get(3)?,
+                        created_at: row.get(4)?,
+                        balance_total: row.get(5).ok(),
+                        last_activity: row.get::<_, Option<i64>>(6).ok().flatten().map(|t| t.to_string()),
+                        status: row.get(7)?,
+                        contact_count: Some(row.get(8)?),
+                        user_id: row.get(9)?,
+                        is_active: row.get::<_, i64>(10).unwrap_or(1) != 0,
+                        balance_fiat: None,
+                        fiat_currency: None,
+                    })
+                },
+            ) {
+                Ok(metadata) => Ok(Some(metadata)),
+                Err(bdk_wallet::rusqlite::Error::QueryReturnedNoRows) => Ok(None),
+                Err(e) => Err(e.into()),
+            }
+        })
+        .await?
+    }
+
+    pub async fn get_wallet_by_checksum(&self, checksum: &str) -> Result<Option<WalletMetadata>> {
+        let pool = self.pool.clone();
+        let checksum = checksum.to_string();
+
+        spawn_blocking(move || -> Result<Option<WalletMetadata>> {
+            let conn = pool.get()?;
+            match conn.query_row(
+                "SELECT w.checksum, w.name, w.descriptor, w.hex_color, w.created_at, w.balance_total,
+                        (SELECT MAX(t.first_seen_at) FROM transactions t WHERE t.wallet_checksum = w.checksum) as last_activity,
+                        w.status, COUNT(c.id) as contact_count, w.user_id, w.is_active
+                 FROM wallets w
+                 LEFT JOIN contacts c ON w.checksum = c.wallet_checksum
+                 WHERE w.checksum = ?1
+                 GROUP BY w.checksum, w.name, w.descriptor, w.hex_color, w.created_at, w.balance_total, w.status, w.user_id, w.is_active",
+                params![checksum],
+                |row| {
+                    Ok(WalletMetadata {
+                        checksum: row.get(0)?,
+                        name: row.get(1)?,
+                        descriptor: row.get(2)?,
+                        hex_color: row.get(3)?,
+                        created_at: row.get(4)?,
+                        balance_total: row.get(5).ok(),
+                        last_activity: row.get::<_, Option<i64>>(6).ok().flatten().map(|t| t.to_string()),
+                        status: row.get(7)?,
+                        contact_count: Some(row.get(8)?),
+                        user_id: row.get(9)?,
+                        is_active: row.get::<_, i64>(10).unwrap_or(1) != 0,
+                        balance_fiat: None,
+                        fiat_currency: None,
+                    })
+                },
+            ) {
+                Ok(metadata) => Ok(Some(metadata)),
+                Err(bdk_wallet::rusqlite::Error::QueryReturnedNoRows) => Ok(None),
+                Err(e) => Err(e.into()),
+            }
+        })
+        .await?
+    }
+
+    pub async fn update_wallet_by_checksum(
+        &self,
+        wallet_checksum: &str,
+        name: &str,
+    ) -> Result<bool> {
+        let pool = self.pool.clone();
+        let name = name.to_string();
+        let checksum = wallet_checksum.to_string();
+
+        spawn_blocking(move || -> Result<bool> {
+            let conn = pool.get()?;
+            let changes = conn.execute(
+                "UPDATE wallets SET name = ?1 WHERE checksum = ?2",
+                params![&name, checksum],
+            )?;
+            Ok(changes > 0)
+        })
+        .await?
+    }
+
+    /// Mark a wallet as deleted (soft delete) - used for non-blocking deletion
+    pub async fn mark_wallet_as_deleted(&self, checksum: &str) -> Result<bool> {
+        let pool = self.pool.clone();
+        let checksum = checksum.to_string();
+
+        spawn_blocking(move || -> Result<bool> {
+            let conn = pool.get()?;
+            let changes = conn.execute(
+                "UPDATE wallets SET status = 'deleted' WHERE checksum = ?1",
+                params![checksum],
+            )?;
+            Ok(changes > 0)
+        })
+        .await?
+    }
+
+    /// Hard delete wallet from database (used after soft delete cleanup)
+    pub async fn hard_delete_wallet_by_checksum(&self, checksum: &str) -> Result<bool> {
+        let pool = self.pool.clone();
+        let checksum = checksum.to_string();
+
+        spawn_blocking(move || -> Result<bool> {
+            let conn = pool.get()?;
+
+            // First delete notification_logs that reference this wallet's transactions
+            // This prevents foreign key constraint failures when transactions are cascade deleted
+            conn.execute(
+                "DELETE FROM notification_logs WHERE transaction_wallet_checksum = ?1",
+                params![checksum],
+            )?;
+
+            // Now delete the wallet - this will cascade delete transactions, contacts, etc.
+            let changes =
+                conn.execute("DELETE FROM wallets WHERE checksum = ?1", params![checksum])?;
+            Ok(changes > 0)
+        })
+        .await?
+    }
+
+    // ============================
+    // WALLET QUERY OPERATIONS
+    // ============================
+
+    pub async fn get_all_wallets(&self) -> Result<Vec<WalletMetadata>> {
+        self.get_wallets_for_user(None).await
+    }
+
+    pub async fn get_wallets_for_user(&self, user_id: Option<&str>) -> Result<Vec<WalletMetadata>> {
+        let pool = self.pool.clone();
+        let user_id = user_id.map(|id| id.to_string());
+
+        spawn_blocking(move || -> Result<Vec<WalletMetadata>> {
+            let conn = pool.get()?;
+
+            let query = match user_id {
+                Some(_) => {
+                    "SELECT w.checksum, w.name, w.descriptor, w.hex_color, w.created_at, w.balance_total,
+                            (SELECT MAX(t.first_seen_at) FROM transactions t WHERE t.wallet_checksum = w.checksum) as last_activity,
+                            w.status, COUNT(c.id) as contact_count, w.user_id, w.is_active
+                     FROM wallets w
+                     LEFT JOIN contacts c ON w.checksum = c.wallet_checksum
+                     WHERE w.user_id = ?1 AND w.status != 'deleted'
+                     GROUP BY w.checksum, w.name, w.descriptor, w.hex_color, w.created_at, w.balance_total, w.status, w.user_id, w.is_active
+                     ORDER BY w.created_at DESC"
+                }
+                None => {
+                    "SELECT w.checksum, w.name, w.descriptor, w.hex_color, w.created_at, w.balance_total,
+                            (SELECT MAX(t.first_seen_at) FROM transactions t WHERE t.wallet_checksum = w.checksum) as last_activity,
+                            w.status, COUNT(c.id) as contact_count, w.user_id, w.is_active
+                     FROM wallets w
+                     LEFT JOIN contacts c ON w.checksum = c.wallet_checksum
+                     WHERE w.status != 'deleted'
+                     GROUP BY w.checksum, w.name, w.descriptor, w.hex_color, w.created_at, w.balance_total, w.status, w.user_id, w.is_active
+                     ORDER BY w.created_at DESC"
+                }
+            };
+
+            let mut stmt = conn.prepare(query)?;
+
+            let params: Vec<&dyn ToSql> = match user_id.as_ref() {
+                Some(uid) => vec![uid],
+                None => vec![],
+            };
+
+            let wallet_iter = stmt.query_map(&params[..], |row| {
+                Ok(WalletMetadata {
+                    checksum: row.get(0)?,
+                    name: row.get(1)?,
+                    descriptor: row.get(2)?,
+                    hex_color: row.get(3)?,
+                    created_at: row.get(4)?,
+                    balance_total: Some(row.get(5).unwrap_or(0)),
+                    last_activity: row
+                        .get::<_, Option<i64>>(6)
+                        .ok()
+                        .flatten()
+                        .map(|t| t.to_string()),
+                    status: row.get(7)?,
+                    contact_count: Some(row.get(8)?),
+                    user_id: row.get(9)?,
+                    is_active: row.get::<_, i64>(10).unwrap_or(1) != 0, // SQLite stores bool as int
+                    balance_fiat: None,
+                    fiat_currency: None,
+                })
+            })?;
+
+            let mut wallets = Vec::new();
+            for wallet in wallet_iter {
+                wallets.push(wallet?);
+            }
+
+            Ok(wallets)
+        })
+        .await?
+    }
+
+    /// Get wallets for a user ordered by creation time (oldest first) for subscription limits enforcement
+    pub async fn get_wallets_for_user_oldest_first(
+        &self,
+        user_id: &str,
+    ) -> Result<Vec<WalletMetadata>> {
+        let pool = self.pool.clone();
+        let user_id = user_id.to_string();
+
+        spawn_blocking(move || -> Result<Vec<WalletMetadata>> {
+            let conn = pool.get()?;
+
+            let query = "SELECT w.checksum, w.name, w.descriptor, w.hex_color, w.created_at, w.balance_total,
+                               (SELECT MAX(t.first_seen_at) FROM transactions t WHERE t.wallet_checksum = w.checksum) as last_activity,
+                               w.status, COUNT(c.id) as contact_count, w.user_id, w.is_active
+                        FROM wallets w
+                        LEFT JOIN contacts c ON w.checksum = c.wallet_checksum
+                        WHERE w.user_id = ?1
+                        GROUP BY w.checksum, w.name, w.descriptor, w.hex_color, w.created_at, w.balance_total, w.status, w.user_id, w.is_active
+                        ORDER BY w.created_at ASC"; // Oldest first for subscription limits
+
+            let mut stmt = conn.prepare(query)?;
+
+            let wallet_iter = stmt.query_map([&user_id], |row| {
+                Ok(WalletMetadata {
+                    checksum: row.get(0)?,
+                    name: row.get(1)?,
+                    descriptor: row.get(2)?,
+                    hex_color: row.get(3)?,
+                    created_at: row.get(4)?,
+                    balance_total: Some(row.get(5).unwrap_or(0)),
+                    last_activity: row
+                        .get::<_, Option<i64>>(6)
+                        .ok()
+                        .flatten()
+                        .map(|t| t.to_string()),
+                    status: row.get(7)?,
+                    contact_count: Some(row.get(8)?),
+                    user_id: row.get(9)?,
+                    is_active: row.get::<_, i64>(10).unwrap_or(1) != 0, // SQLite stores bool as int
+                    balance_fiat: None,
+                    fiat_currency: None,
+                })
+            })?;
+
+            let mut wallets = Vec::new();
+            for wallet in wallet_iter {
+                wallets.push(wallet?);
+            }
+
+            Ok(wallets)
+        })
+        .await?
+    }
+
+    pub async fn count_wallets_for_user(&self, user_id: &str) -> Result<usize> {
+        let pool = self.pool.clone();
+        let user_id = user_id.to_string();
+
+        spawn_blocking(move || -> Result<usize> {
+            let conn = pool.get()?;
+            let count: i64 = conn.query_row(
+                "SELECT COUNT(*) FROM wallets WHERE user_id = ?1",
+                params![user_id],
+                |row| row.get(0),
+            )?;
+            Ok(count as usize)
+        })
+        .await?
+    }
+
+    pub async fn is_wallet_owned_by_user(
+        &self,
+        wallet_checksum: &str,
+        user_id: &str,
+    ) -> Result<bool> {
+        let pool = self.pool.clone();
+        let checksum = wallet_checksum.to_string();
+        let user_id = user_id.to_string();
+
+        spawn_blocking(move || -> Result<bool> {
+            let conn = pool.get()?;
+            let exists: bool = conn
+                .prepare("SELECT 1 FROM wallets WHERE checksum = ?1 AND user_id = ?2")?
+                .exists(params![checksum, user_id])?;
+            Ok(exists)
+        })
+        .await?
+    }
+
+    /// Get all wallets for a specific tier that are due for sync
+    pub async fn get_wallets_for_tier_sync(
+        &self,
+        tier: &crate::subscription::SubscriptionTier,
+        network: &crate::config::NetworkConfig,
+    ) -> Result<Vec<WalletMetadata>> {
+        let pool = self.pool.clone();
+        let network = network.clone();
+        let tier_str = match tier {
+            crate::subscription::SubscriptionTier::Personal => "personal",
+            crate::subscription::SubscriptionTier::Team => "team",
+        }
+        .to_string();
+
+        spawn_blocking(move || -> Result<Vec<WalletMetadata>> {
+            let conn = pool.get()?;
+            let tier_limits =
+                crate::subscription::SubscriptionTier::from(tier_str.clone()).limits(&network);
+
+            let mut stmt = conn.prepare(
+                "SELECT w.checksum, w.name, w.descriptor, w.hex_color, w.balance_total,
+                        w.last_activity, w.last_synced_at, w.status, w.user_id, w.created_at
+                 FROM wallets w
+                 JOIN users u ON w.user_id = u.id
+                 WHERE w.is_active = 1 AND w.status = 'ready'
+                   AND u.subscription_tier = ?1
+                   AND (
+                    -- Admin users bypass all subscription checks
+                    u.is_admin = 1
+                    OR
+                    -- Regular users need valid subscription
+                    (
+                      -- Active subscriptions
+                      u.subscription_status = 'active'
+                      OR
+                      -- Trial users within trial period
+                      (u.subscription_status = 'trialing' AND datetime(u.trial_ends_at) > datetime('now'))
+                      OR
+                      -- Cancelled users still within their paid period
+                      (u.subscription_status = 'canceled' AND u.subscription_ends_at IS NOT NULL AND datetime(u.subscription_ends_at) > datetime('now'))
+                    )
+                 )
+                 AND (
+                    -- Admin users bypass timing restrictions
+                    u.is_admin = 1
+                    OR
+                    -- Regular users follow timing rules
+                    (
+                      -- Never synced before
+                      w.last_synced_at IS NULL
+                      OR
+                      -- Or due for sync based on tier interval
+                      datetime(w.last_synced_at) <= datetime('now', '-' || ?2 || ' seconds')
+                    )
+                 )
+                 ORDER BY w.checksum",
+            )?;
+
+            let wallet_rows =
+                stmt.query_map(params![tier_str, tier_limits.sync_interval_secs], |row| {
+                    Ok(WalletMetadata {
+                        checksum: row.get(0)?,
+                        name: row.get(1)?,
+                        descriptor: row.get(2)?,
+                        hex_color: row.get(3)?,
+                        balance_total: row.get(4)?,
+                        last_activity: row.get(5)?,
+                        status: row.get(7)?,
+                        contact_count: None, // Not counting contacts in this query
+                        user_id: row.get(8)?,
+                        created_at: row.get(9)?,
+                        is_active: true, // Query already filters for active wallets
+                        balance_fiat: None,
+                        fiat_currency: None,
+                    })
+                })?;
+
+            let mut due_wallets = Vec::new();
+            for row in wallet_rows {
+                due_wallets.push(row?);
+            }
+
+            Ok(due_wallets)
+        })
+        .await?
+    }
+
+    /// Get summary of wallets that are not being synced due to subscription issues
+    pub async fn get_non_syncing_wallets_summary(&self) -> Result<NonSyncingWalletsSummary> {
+        let pool = self.pool.clone();
+
+        spawn_blocking(move || -> Result<NonSyncingWalletsSummary> {
+            let conn = pool.get()?;
+
+            // Count expired trials (trialing but past trial_ends_at)
+            let expired_trials: usize = conn.query_row(
+                "SELECT COUNT(DISTINCT w.checksum)
+                 FROM wallets w
+                 JOIN users u ON w.user_id = u.id
+                 WHERE w.is_active = 1 AND w.status = 'ready'
+                   AND u.is_admin = 0
+                   AND u.subscription_status = 'trialing'
+                   AND datetime(u.trial_ends_at) <= datetime('now')",
+                [],
+                |row| row.get(0),
+            )?;
+
+            // Count cancelled subscriptions (cancelled and past subscription_ends_at)
+            let cancelled_subscriptions: usize = conn.query_row(
+                "SELECT COUNT(DISTINCT w.checksum)
+                 FROM wallets w
+                 JOIN users u ON w.user_id = u.id
+                 WHERE w.is_active = 1 AND w.status = 'ready'
+                   AND u.is_admin = 0
+                   AND u.subscription_status = 'canceled'
+                   AND (u.subscription_ends_at IS NULL OR datetime(u.subscription_ends_at) <= datetime('now'))",
+                [],
+                |row| row.get(0),
+            )?;
+
+            // Count expired subscriptions
+            let expired_subscriptions: usize = conn.query_row(
+                "SELECT COUNT(DISTINCT w.checksum)
+                 FROM wallets w
+                 JOIN users u ON w.user_id = u.id
+                 WHERE w.is_active = 1 AND w.status = 'ready'
+                   AND u.is_admin = 0
+                   AND u.subscription_status = 'expired'",
+                [],
+                |row| row.get(0),
+            )?;
+
+            // Count past_due subscriptions
+            let past_due_subscriptions: usize = conn.query_row(
+                "SELECT COUNT(DISTINCT w.checksum)
+                 FROM wallets w
+                 JOIN users u ON w.user_id = u.id
+                 WHERE w.is_active = 1 AND w.status = 'ready'
+                   AND u.is_admin = 0
+                   AND u.subscription_status = 'past_due'",
+                [],
+                |row| row.get(0),
+            )?;
+
+            // Count inactive wallets (due to tier limits)
+            let inactive_wallets: usize = conn.query_row(
+                "SELECT COUNT(*)
+                 FROM wallets w
+                 WHERE w.is_active = 0 AND w.status = 'ready'",
+                [],
+                |row| row.get(0),
+            )?;
+
+            let total_non_syncing = expired_trials
+                + cancelled_subscriptions
+                + expired_subscriptions
+                + past_due_subscriptions
+                + inactive_wallets;
+
+            Ok(NonSyncingWalletsSummary {
+                expired_trials,
+                cancelled_subscriptions,
+                expired_subscriptions,
+                past_due_subscriptions,
+                inactive_wallets,
+                total_non_syncing,
+            })
+        })
+        .await?
+    }
+
+    pub async fn get_ready_wallets(&self) -> Result<Vec<WalletMetadata>> {
+        let pool = self.pool.clone();
+
+        spawn_blocking(move || -> Result<Vec<WalletMetadata>> {
+            let conn = pool.get()?;
+
+            let query = "SELECT w.checksum, w.name, w.descriptor, w.hex_color,
+                                w.created_at, w.balance_total,
+                                (SELECT MAX(t.first_seen_at) FROM transactions t
+                                 WHERE t.wallet_checksum = w.checksum) as last_activity,
+                                w.status, COUNT(c.id) as contact_count,
+                                w.user_id, w.is_active
+                         FROM wallets w
+                         LEFT JOIN contacts c ON w.checksum = c.wallet_checksum
+                         WHERE w.status = 'ready'
+                         GROUP BY w.checksum, w.name, w.descriptor, w.hex_color,
+                                  w.created_at, w.balance_total, w.status,
+                                  w.user_id, w.is_active
+                         ORDER BY w.created_at DESC";
+
+            let mut stmt = conn.prepare(query)?;
+
+            let wallet_iter = stmt.query_map([], |row| {
+                Ok(WalletMetadata {
+                    checksum: row.get(0)?,
+                    name: row.get(1)?,
+                    descriptor: row.get(2)?,
+                    hex_color: row.get(3)?,
+                    created_at: row.get(4)?,
+                    balance_total: Some(row.get(5).unwrap_or(0)),
+                    last_activity: row
+                        .get::<_, Option<i64>>(6)
+                        .ok()
+                        .flatten()
+                        .map(|t| t.to_string()),
+                    status: row.get(7)?,
+                    contact_count: row.get(8).unwrap_or(Some(0)),
+                    user_id: row.get(9)?,
+                    is_active: row.get(10)?,
+                    balance_fiat: None,
+                    fiat_currency: None,
+                })
+            })?;
+
+            wallet_iter
+                .collect::<std::result::Result<Vec<_>, _>>()
+                .map_err(|e| anyhow!("Failed to query ready wallets: {}", e))
+        })
+        .await?
+    }
+
+    /// Get all wallets marked as deleted
+    pub async fn get_deleted_wallets(&self) -> Result<Vec<WalletMetadata>> {
+        let pool = self.pool.clone();
+
+        spawn_blocking(move || -> Result<Vec<WalletMetadata>> {
+            let conn = pool.get()?;
+
+            let query = "SELECT w.checksum, w.name, w.descriptor, w.hex_color,
+                                w.created_at, w.balance_total,
+                                (SELECT MAX(t.first_seen_at) FROM transactions t
+                                 WHERE t.wallet_checksum = w.checksum) as last_activity,
+                                w.status, COUNT(c.id) as contact_count,
+                                w.user_id, w.is_active
+                         FROM wallets w
+                         LEFT JOIN contacts c ON w.checksum = c.wallet_checksum
+                         WHERE w.status = 'deleted'
+                         GROUP BY w.checksum, w.name, w.descriptor, w.hex_color,
+                                  w.created_at, w.balance_total, w.status,
+                                  w.user_id, w.is_active
+                         ORDER BY w.created_at DESC";
+
+            let mut stmt = conn.prepare(query)?;
+
+            let wallet_iter = stmt.query_map([], |row| {
+                Ok(WalletMetadata {
+                    checksum: row.get(0)?,
+                    name: row.get(1)?,
+                    descriptor: row.get(2)?,
+                    hex_color: row.get(3)?,
+                    created_at: row.get(4)?,
+                    balance_total: Some(row.get(5).unwrap_or(0)),
+                    last_activity: row
+                        .get::<_, Option<i64>>(6)
+                        .ok()
+                        .flatten()
+                        .map(|t| t.to_string()),
+                    status: row.get(7)?,
+                    contact_count: row.get(8).unwrap_or(Some(0)),
+                    user_id: row.get(9)?,
+                    is_active: row.get(10)?,
+                    balance_fiat: None,
+                    fiat_currency: None,
+                })
+            })?;
+
+            wallet_iter
+                .collect::<std::result::Result<Vec<_>, _>>()
+                .map_err(|e| anyhow!("Failed to query deleted wallets: {}", e))
+        })
+        .await?
+    }
+
+    // ============================
+    // WALLET STATUS & SYNC OPERATIONS
+    // ============================
+
+    pub async fn update_wallet_balance_by_checksum(
+        &self,
+        wallet_checksum: &str,
+        balance_total: i64,
+    ) -> Result<()> {
+        let pool = self.pool.clone();
+        let checksum = wallet_checksum.to_string();
+
+        spawn_blocking(move || -> Result<()> {
+            let conn = pool.get()?;
+            conn.execute(
+                "UPDATE wallets SET balance_total = ?1 WHERE checksum = ?2",
+                params![balance_total, checksum],
+            )?;
+            Ok(())
+        })
+        .await?
+    }
+
+    pub async fn update_wallet_last_synced(&self, checksum: &str) -> Result<()> {
+        let pool = self.pool.clone();
+        let checksum = checksum.to_string();
+
+        spawn_blocking(move || -> Result<()> {
+            let conn = pool.get()?;
+            let current_time = chrono::Utc::now()
+                .format("%Y-%m-%d %H:%M:%S%.3f")
+                .to_string();
+            conn.execute(
+                "UPDATE wallets SET last_synced_at = ?1 WHERE checksum = ?2",
+                params![current_time, checksum],
+            )?;
+            Ok(())
+        })
+        .await?
+    }
+
+    pub async fn update_wallet_status(&self, checksum: &str, status: &str) -> Result<()> {
+        let pool = self.pool.clone();
+        let checksum = checksum.to_string();
+        let status = status.to_string();
+
+        spawn_blocking(move || -> Result<()> {
+            let conn = pool.get()?;
+            conn.execute(
+                "UPDATE wallets SET status = ?1 WHERE checksum = ?2",
+                params![status, checksum],
+            )?;
+            Ok(())
+        })
+        .await?
+    }
+
+    pub async fn update_wallet_active_status(&self, checksum: &str, is_active: bool) -> Result<()> {
+        let pool = self.pool.clone();
+        let checksum = checksum.to_string();
+
+        tokio::task::spawn_blocking(move || {
+            let conn = pool.get()?;
+
+            conn.execute(
+                "UPDATE wallets SET is_active = ? WHERE checksum = ?",
+                params![is_active, checksum],
+            )?;
+
+            Ok::<(), anyhow::Error>(())
+        })
+        .await??;
+
+        Ok(())
+    }
+}
